+++ src/libigl/igl/SortableRow.h
@@ -47,7 +47,7 @@ public:
         {
             return false;
         }
-        for(int i = 0;i<this->data.size();i++)
+        for( size_t i = 0;i<this->data.size();i++)
         {
             if(this->data(i) != that.data(i))
             {
